add login page

+++ Models/Forecast.cs
@@ -10,7 +10,7 @@ namespace MyPlayground.Models
         }
 
         public string CurrentDateTimeFormatted { get; set; }
-        public double CurrentTemperatureC { get; set; }
+        public int CurrentTemperatureC { get; set; }
         public double CurrentHumidity { get; set; }
         public double CurrentWindSpeed { get; set; }
         public string CurrentSummary { get; set; }
